// src/models/avanzaRecord.ts
export class AvanzaRecord {
    date: Date;
    account: string;
    type: string;
    description: string;  
    quantity: number;
    price: number;
    amount: number;
    currency: string;
    fee: number;
    exchangeRate: number;
    instrumentCurrency: string;
    isin: string;
    result: number; 
}